refactor forms and autocomplete view
