+++ dashboard/urls.py
@@ -51,11 +51,8 @@ urlpatterns = [
     url(r'^product_puc/(?P<pk>\d+)$', views.assign_puc_to_product,
                                             name='product_puc'),
     url(r'^puc-autocomplete/$',
-                    views.product_autocomplete.PUCAutocomplete.as_view(),
+                    views.puc_autocomplete.PUCAutocomplete.as_view(),
                                             name='puc-autocomplete'),
-    url(r'^habit-autocomplete/$',
-                    views.habit_autocomplete.HabitAutocomplete.as_view(),
-                                            name='habit-autocomplete'),
     url(r'^product/(?P<pk>\d+)$', views.product_detail,
                                             name='product_detail'),
     url(r'^products/$', views.product_list, name='product_list'),
