+++ dashboard/views/puc_autocomplete.py
@@ -3,7 +3,7 @@ from dashboard.models import PUC
 from django.db.models import Q
 
 
-class HabitAutocomplete(autocomplete.Select2QuerySetView):
+class PUCAutocomplete(autocomplete.Select2QuerySetView):
     def get_queryset(self):
         # Don't forget to filter out results depending on the visitor !
         if not self.request.user.is_authenticated:
